Rename source files whose have same name as its parent folder to avoid unnecessary nesting in imports. !2
